requirements.txt
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
Werkzeug>=3.1.3
Flask>=2.3.3
python-dotenv>=1.1.1